Conftool: initial commit
conftool is a python library that allows interacting with a
filesystem-like KV store via a hierarchical, semi-schema-based object
representation. on top of that, two tools are being built:

- conf-syncer which allows to keep data in the kv store up to date with
  our on-disk representation. This is a bit too much specific to our
  current use-case right now.
- conftool (still not committed) that will allow to interact from the
  command line with specific entities in our datastore.

A good part of what is here now is still work in progress. The best
feature is that all the backend-specific code is constrained into a
driver class, so that it should be easy to switch backends in the future
if we ever decided to.

Change-Id: I199e24c3e7a959e8a4156149c6213ab1c9199b32