dbctl: validate the instance given to section set-master
Previously if you tried to section FOO set-master to an instance that
didn't exist, or to an instance that was not a replica of that section,
you'd get a confusing error message, something like:
    Section s1 is supposed to have master db1178 but had db1067 instead
due to some implementation details of compute_config.

So instead, check up-front for some preconditions.  Now we get errors like:
    DB instance 'dba1' is not a replica of section 's2'

Change-Id: Ib110f832505befe7898013150650032c69d90f77