dbctl: always validate vs JSON schema
Previously, the only time that we validated the computed config against
the JSON schema was immediately before writing to etcd, after showing a
diff and prompting the user for a message & confirmation.  That's pretty
silly -- while it should almost never happen that we fail to validate
something we computed against our own schema, it can happen in the case of
a bug.  So we should whine early and often in such an occurrence.

Discovered this as part of adding a new 'hostsByName' key to the top-level
config object.

Change-Id: Ib78d709cf7fb0d2b0a94c05267de12733c2f3eaf