Fix configuration file lookup when running with sudo
On systems with env_keep += HOME, which is quite common, confctl would
look for '~' via os.path.expandpath, which defaults to the $HOME env
var. So 'sudo -u foo confctl' would, very confusingly, not load the
.etcdrc file from the foo user, but the one of the user launching the
command.

Also suppress the annoying messages for missing yaml files when not
running in debug mode.

Change-Id: I1bc0daf3d3926b8f8348641045de595c352b1827