check return value of $EDITOR in EditActions
Affects confctl edit and dbctl instance/section edit.

Means that you can bail out of an edit with :cq in vim,
and maybe in Emacs with C-u 1 M-x kill-emacs (untested).
Sending SIGKILL to your editor should also work.

Something that should also be added: if there's no real delta
before and after the editor invocation, we should probably not
write anything to etcd (except if overridden with a flag; this
is handy for schema changes).

Bug: T233680
Change-Id: Iad329f267f01e0ad289a4894bd0daaf9c76b53a7