dbctl test fixtures: one schema to rule them all
Make the schema used in dbctl testing as close as possible to the one
used in production.

(There's one difference from current production, which is that dnsdisc
no longer has a dependency on service objects, as service objects no
longer exist.)

This not only improves testing but also simplifies a very annoying step
in doing releases: looking at the schema in Puppet and looking at the
schema here and trying to rationalize the expected differences vs the
unexpected differences.  Now the only expected differences are the
necessarily-different values for base_path.  (In theory we could modify
the testing environment to make those unnecessary?)

Change-Id: I76e1bb6ed9d0ec9152d69d48584327385be60a6f