requestctl: set an X-Requestctl header for matching rules
The VCL we produce will now generate a "X-Requestctl" header which
will contain all the enabled rules that match the specific request.

We can then add VCL fragments to the varnish configuration to handle this
header and mold it to anything we might want.

The next step is to add an additional boolean switch to the schema for action
so that we're allowed to add also actions in log_only mode.

Bug: T305582
Change-Id: Ib887d259f1a920af702a638110221db14c426af7