requestctl: add AND NOT and OR NOT to the parsing grammar
Also: refactor how we do translations from our DSL to VCL/VSL.

The general structure of the two varnish-related DSLs is quite similar
and generated a lot of duplicated logic. Instead, create a translator
interface and specific classes for translating expressions to VCL and
VSL.

The advantage of this approach is that we can progressively generalize
it when we start wanting to inject requestctl rules at other layers of
our request stack.

Bug: T305607
Change-Id: I9d78a4ff680579fd6c7754f9cff09671a3d14204