requestctl: add retry-after request header when applicable
When throttling requests, it makes sense to set Retry-After in the response
to a meaningful value, corresponding to the throttle duration.

To this end, add a Retry-After *request* header with a value equal to the
throttle duration or at least 1 second.

We will have to extract the value of this header in the VCL to use it as a
response header in the synth phase.

Bug: T305824
Change-Id: Ice339d9b390aa85ab11f834bdc37480a4495a9cf