dbctl: extend dbconfig checks to external sections
This change extends the existing suite of dbconfig validity checks to
cover external-flavored sections as well.

This includes both the limited suite of checks on restore, as well as
the more complete checks applied on config updates.

A handful of older test fixtures missing externalLoads have been updated
to include it (same as valid.json).

Bug: T365123
Change-Id: I65e7c4125769d669c8fa300e869322371913b79a