dbctl: break up test_check_config test case
This was a bit long, which was exacerbated by adding coverage for the
external-flavor case (~ 2x the number of assertions). Tests are now 1:1
with scenario and asserted behavior.

While parameterization could make this more concise, doing so while
expressing test scenarios in terms of mutations to config or fixture
gets complicated quickly.

Note that test cases are slightly reordered, so that regular and
external flavor versions appear in the same order.

Bug: T365123
Change-Id: I6daf5800d0857dbc097173d0fdc3331ff66ffeec