Release 3.0.0
While nearly all changes since 2.3.3 are trivially backward compatible,
extending dbconfig checks to external sections is less clear.

Even though that's a breaking change *only* in the presence of what is
arguably broken configuration, it's really the observable behavior of
the tool that defines the API (not intent).

Thus, treating the validation change as warranting a major verison bump
seems like the only "technically correct" choice.

Bug: T365123
Change-Id: Ic3321c97e439bc11511a5f70e17b0e525c600d3d