drivers/etcd: only attempt to load existing configs
In an earlier change, yaml_safe_load was made to error-log when the
provided file cannot be loaded, including the case where it doesn't
exist.

While that's probably a good thing in the general case, it may be a bit
too chatty for the etcdrc file "search" in get_config.

Thus, rather than special-casing does-not-exist in yaml_safe_load, it's
probably clearer to just avoid loading non-existent configs.

Bug: T367919
Change-Id: I181485227b66757a3e3ce6b67df53022122afca7